Moved adapters to respective packages
