+++ app/src/main/java/com/acmvit/acm_app/model/Idea.java
@@ -1,4 +1,4 @@
-package com.acmvit.acm_app.ui.ideas;
+package com.acmvit.acm_app.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
