+++ app/src/main/java/com/acmvit/acm_app/ui/ideas/DashboardFragment.java
@@ -11,10 +11,12 @@ import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 import androidx.viewpager2.widget.ViewPager2;
-import androidx.viewpager2.widget.ViewPager2;
+
 import com.acmvit.acm_app.R;
+import com.acmvit.acm_app.model.Idea;
+import com.acmvit.acm_app.ui.ideas.adapters.DashboardViewPagerAdapter;
+
 import java.util.ArrayList;
 
 public class DashboardFragment extends Fragment {
