+++ app/src/main/java/com/acmvit/acm_app/ui/members/MembersFragment.java
@@ -5,12 +5,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.SearchView;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
-import com.acmvit.acm_app.MembersAdapter;
+
 import com.acmvit.acm_app.databinding.FragmentMembersBinding;
+import com.acmvit.acm_app.ui.members.adapters.MembersAdapter;
+
 import org.jetbrains.annotations.NotNull;
 
 public class MembersFragment extends Fragment {
