+++ app/src/main/java/com/acmvit/acm_app/ui/splash/WelcomeActivity.java
@@ -11,6 +11,7 @@ import androidx.core.content.ContextCompat;
 import androidx.viewpager.widget.ViewPager;
 import com.acmvit.acm_app.R;
 import com.acmvit.acm_app.ui.MainActivity;
+import com.acmvit.acm_app.ui.splash.adapters.WelcomeAdapter;
 
 public class WelcomeActivity extends AppCompatActivity {
